User/rehan

+++ C++/src/math/uglynumber2.cpp
@@ -48,7 +48,7 @@ int nthUglyNumber(int n) {
             
             if (count == n)
             {
-                rv = i;
+                rv = (int)i;
                 return rv;
             }
         }
